speed up checks with caching and build matrices
This commit refactors the GitHub Actions workflow by further
parallelizing jobs with build matrices and caching the artifacts.

The cache is keyed off of Cargo.lock, but includes a partial key restore
so a cache is still used even when Cargo.lock changes. It will then
create a new cache for future builds of the same Cargo.lock.

A file named cache_bust is added to the .github folder. Any edit to that
file will cause a complete cache miss. We can use this to start over if
we ever run into trouble.